moving AuthenticationInfo to org.apache.zeppelin.display.AuthenticationInfo
