+++ zeppelin-display/src/test/scala/org/apache/zeppelin/display/angular/AbstractAngularModelTest.scala
@@ -16,7 +16,7 @@
  */
 package org.apache.zeppelin.display.angular
 
-import org.apache.zeppelin.display.{AngularObjectRegistry, GUI}
+import org.apache.zeppelin.display.{AuthenticationInfo, AngularObjectRegistry, GUI}
 import org.apache.zeppelin.interpreter._
 import org.scalatest.concurrent.Eventually
 import org.scalatest.{BeforeAndAfter, BeforeAndAfterEach, FlatSpec, Matchers}
